Filtration of the "ghost mushrooms", making modyfied copies of the mushrooms images and creating second model
I deleted those mushrooms that didnt have its corresponding id in the name_classification.csv, which made them unable to classify to any of the classes we needed to keep/delete.
To ensure that these mushrooms were not needed i checked their names and left 35 of them in our dataset.
Furthermore i modified and added some more image transformations and implemented a function to take that action back if accidently played twice.
At last I modified the InceptionV3 model to fit in our current data and trained it for 10 epochs